requirements.txt
aioredis==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
bcrypt==4.2.1
certifi==2024.8.30
click==8.1.7
distro==1.9.0
ecdsa==0.19.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
Mako==1.3.8
MarkupSafe==3.0.2
openai==1.57.4
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.3.4
pytest-asyncio==0.25.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
rsa==4.9
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.33.0